Implement Flask blueprints and application factory
