+++ CHANGELOG.md
@@ -4,6 +4,14 @@
 
 ### Changed
 
+- A file named **.golem** must exist in the test directory
+    
+    Migrate from previous versions: Create a file named **.golem** in the test directory root with the following content:
+    ```
+    [gui]
+    secret_key = your_secret_key_string
+    ```
+
 - golem.execution.workspace -> golem.execution.testdir
