+++ golem/core/errors.py
@@ -1,0 +1,3 @@
+
+invalid_test_directory = ('Error: {} is not an valid Golem test directory; '
+                          '.golem file not found')
