+++ tests/gui/gui_utils_test.py
@@ -1,6 +1,8 @@
+import os
 import pytest
 
 from golem.gui import gui_utils
+from golem.core import errors
 
 
 DOCSTRINGS = [
@@ -86,3 +88,41 @@ class TestGolemActionParser:
     def test__parse_docstring(self, actual, expected):
         expected = gui_utils.Golem_action_parser()._parse_docstring(actual)
         assert expected == expected
+
+
+class TestGetSecretKey:
+
+    def test_get_secret_key(self, testdir_function):
+        testdir_function.activate()
+        secret_key = gui_utils.get_secret_key()
+        assert type(secret_key) == str
+
+    def test_get_secret_key_golem_file_missing(self, testdir_function):
+        testdir = testdir_function.activate()
+        golem_file_path = os.path.join(testdir, '.golem')
+        os.remove(golem_file_path)
+        with pytest.raises(SystemExit) as wrapped_execution:
+            gui_utils.get_secret_key()
+        expected = errors.invalid_test_directory.format(testdir)
+        assert wrapped_execution.value.code == expected
+
+    def test_get_secret_key_golem_file_missing_gui_key_section(self, testdir_function, capsys):
+        testdir = testdir_function.activate()
+        golem_file_path = os.path.join(testdir, '.golem')
+        with open(golem_file_path, 'w') as f:
+            f.write('')
+        secret_key = gui_utils.get_secret_key()
+        out, _ = capsys.readouterr()
+        assert 'Warning: gui config section not found in .golem file. Using default secret key' in out
+        assert secret_key == gui_utils.DEFAULT_SECRET_KEY
+
+    def test_get_secret_key_golem_file_missing_secret_key(self, testdir_function, capsys):
+        testdir = testdir_function.activate()
+        golem_file_path = os.path.join(testdir, '.golem')
+        with open(golem_file_path, 'w') as f:
+            f.write('[gui]\n')
+        secret_key = gui_utils.get_secret_key()
+        out, _ = capsys.readouterr()
+        assert 'Warning: secret_key not found in .golem file. Using default secret key' in out
+        assert secret_key == gui_utils.DEFAULT_SECRET_KEY
+
